pyproject.toml
[project]
name = "ingredient-recipe-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "bina",email = "dhtpwjd26122@gmail.com"}
]
license = {text = "MIN"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.123.2,<0.124.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "google-genai (>=1.52.0,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
